+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 This CHANGELOG.md file will contain the update log for the latest set of updates to the templates
 
+# UPDATES for Azure Stack Release 3.9 - April 8, 2018
+
+1.  Enable option for CNS
+2.  With CNS, deploy 3 additional nodes with 3 disks per node for Gluster
+3.  Deploy Service Catalog, Ansible Service Broker and Template Service Broker if CNS is enabled
+
+
 # UPDATES for Azure Stack Release 3.9 - April 4, 2018
 
 1.  Create Azure Stack Release 3.9 Branch